feat: user profile preview
